Create temporary MANIFEST.in file
Python's packaging tools require an ugly MANIFEST.in file to include the
bundled non-Python files in this package. A manifest.in file works on
OSX and Windows (which have case-insensitive file systems) but not on
Linux.

This commit removes the manifest.in file and replaces it with a
mechanism for creating and removing a temporary MANIFEST.in file when
running setup.py.